handling get_longitudinal position function
