+++ example_two_car_lane_switch4.py
@@ -48,7 +48,7 @@ def controller(ego:State, other:State, lane_map:LaneMap):
 
     return output
 
-from src.example.example_agent.car_agent2 import CarAgent2
+from src.example.example_agent.car_agent import CarAgent
 from src.scene_verifier.scenario.scenario import Scenario
 from src.example.example_map.simple_map2 import SimpleMap5, SimpleMap6
 from src.plotter.plotter2D import *
@@ -61,17 +61,17 @@ if __name__ == "__main__":
     input_code_name = 'example_two_car_lane_switch4.py'
     scenario = Scenario()
 
-    car = CarAgent2('car1', file_name=input_code_name)
+    car = CarAgent('car1', file_name=input_code_name)
     scenario.add_agent(car)
-    car = CarAgent2('car2', file_name=input_code_name)
+    car = CarAgent('car2', file_name=input_code_name)
     scenario.add_agent(car)
-    tmp_map = SimpleMap5()
+    tmp_map = SimpleMap6()
     scenario.add_map(tmp_map)
     scenario.set_sensor(FakeSensor2())
     scenario.set_init(
         [
-            [[15, 0, 0, 0.5],[15, 0, 0, 0.5]], 
-            [[0, -0.2, 0, 1.0],[0.2, 0.2, 0, 1.0]],
+            [[14, 0, 0, 0.5],[14, 0, 0, 0.5]], 
+            [[0, -0.2, 0, 1.0],[0.1, 0.2, 0, 1.0]],
         ],
         [
             (VehicleMode.Normal, LaneMode.Lane1),
